refactor: consume markdown types from package
