+++ packages/docops/tsconfig.json
@@ -9,11 +9,10 @@
     "paths": {
       "@shared/ts/*": ["../shared/ts/dist/*"],
       "@promethean/test-utils/*": ["../packages/test-utils/dist/*"],
-      "@promethean/legacy/*": ["../packages/legacy/*"],
-      "@promethean/markdown/*": ["../markdown/dist/*"]
+      "@promethean/legacy/*": ["../packages/legacy/*"]
     }
   },
-  "include": ["src/**/*", "./types/**/*.d.ts"],
+  "include": ["src/**/*"],
   "references": [
     {
       "path": "../test-utils"
