+++ packages/markdown/src/index.ts
@@ -1,0 +1,2 @@
+export { sentenceSplit, parseMarkdownChunks } from './chunking.js';
+export type { MarkdownChunk } from './types.js';
